Former-commit-id: 674e8929ebe73a457e764f82ba49ff659d4fea61

+++ TileCalorimeter/TileRec/TileRec/TileTBHitToNtuple.h
@@ -52,7 +52,7 @@ private:
     std::string m_ntupleLoc;
 
     NTuple::Item<int> m_nchan;
-    NTuple::Item<float> m_tolE;
+    NTuple::Item<double> m_tolE;
 
     NTuple::Array<float> m_energy;
     NTuple::Array<float> m_time;
